Demo Site out for review with client
Most changes were made in realtime online.  I updated my git folder but there was a problem so I had to remove the old repository with all my different changes and create a new one with the completed site.  I still need to work on the NavWalker navigation bar though.